chore: move service's builder to builder
